Move RpcSubscriptions into its own module
