+++ src/lib.rs
@@ -62,6 +62,7 @@ pub mod rpc;
 pub mod rpc_mock;
 pub mod rpc_pubsub;
 pub mod rpc_request;
+pub mod rpc_subscriptions;
 pub mod service;
 pub mod sigverify;
 pub mod sigverify_stage;
